+++ app/application/src/main/kotlin/br/com/fiap/gerenciamentopedidos/application/responses/ProdutoResponse.kt
@@ -2,13 +2,14 @@ package br.com.fiap.gerenciamentopedidos.application.responses
 
 import br.com.fiap.gerenciamentopedidos.domain.dtos.ProdutoDto
 import br.com.fiap.gerenciamentopedidos.domain.enums.Categoria
+import java.math.BigDecimal
 
 class ProdutoResponse(produto: ProdutoDto) {
     var id: Long? = null
     var nome: String? = null
     var descricao: String? = null
     var categoria: Categoria? = null
-    var valor: Double? = null
+    var valor: BigDecimal? = null
     var tempoPreparo: Long? = null
     var disponivel: Boolean? = null
     var excluido: Boolean? = null
